requirements.txt
nltk
translate
python-telegram-bot==13.2
setuptools-rust
